% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multe_data.R
\name{read_multe_data}
\alias{read_multe_data}
\title{Read cloudberry data
Reads the built-in cloudberry dataset from the package.}
\usage{
read_multe_data()
}
\value{
A data frame with columns 'site' and 'seeds'
}
\description{
Read cloudberry data
Reads the built-in cloudberry dataset from the package.
}
\examples{
read_multe_data()
}
