% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Runs the an interactive Shiny app}
\usage{
run_app()
}
\value{
App where you can explore and compare number of seeds in cloudberries
}
\description{
Runs the an interactive Shiny app
}
